+++ src/main/java/kitchenpos/order/domain/delivery/port/RiderPort.java
@@ -1,0 +1,8 @@
+package kitchenpos.order.domain.delivery.port;
+
+import kitchenpos.order.domain.order.model.Order;
+
+public interface RiderPort {
+
+    void requestRider(Order order);
+}
