+++ src/main/java/kitchenpos/order/domain/eatin/flow/EatInOrderFlow.java
@@ -1,0 +1,4 @@
+package kitchenpos.order.domain.eatin.flow;
+
+public class EatInOrderFlow {
+}
