--- src/main/java/kitchenpos/order/domain/model/OrderType.java
@@ -1,5 +1,0 @@
-package kitchenpos.order.domain.model;
-
-public enum OrderType {
-    DELIVERY, TAKEOUT, EAT_IN
-}
