+++ src/main/java/kitchenpos/order/domain/order/model/CustomOrderType.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.domain.order.model;
+
+public interface CustomOrderType {
+    OrderFlow getOrderFlow();
+}
