+++ src/main/java/kitchenpos/order/domain/order/model/OrderFlow.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.domain.order.model;
+
+public interface OrderFlow {
+    OrderStatus performNextFlow(OrderStatus currentStatus);
+}
