--- src/main/java/kitchenpos/order/domain/port/RiderPort.java
@@ -1,8 +1,0 @@
-package kitchenpos.order.domain.port;
-
-import kitchenpos.order.domain.model.Order;
-
-public interface RiderPort {
-
-    void requestRider(Order order);
-}
