--- src/main/java/kitchenpos/order/domain/repository/OrderRepository.java
@@ -1,19 +1,0 @@
-package kitchenpos.order.domain.repository;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-import kitchenpos.order.domain.model.Order;
-import kitchenpos.order.domain.model.OrderStatus;
-import kitchenpos.order.domain.model.OrderTable;
-
-public interface OrderRepository {
-    Order save(Order order);
-
-    Optional<Order> findById(UUID id);
-
-    List<Order> findAll();
-
-    boolean existsByOrderTableAndStatusNot(OrderTable orderTable, OrderStatus status);
-}
-
