+++ src/main/java/kitchenpos/order/domain/takeout/flow/TakeOutOrderFlow.java
@@ -1,0 +1,4 @@
+package kitchenpos.order.domain.takeout.flow;
+
+public class TakeOutOrderFlow {
+}
