+++ src/main/java/kitchenpos/order/domain/takeout/model/TakeOutOrder.java
@@ -1,0 +1,4 @@
+package kitchenpos.order.domain.takeout.model;
+
+public class TakeOutOrder {
+}
