+++ src/main/java/kitchenpos/order/infrastructure/adaptor/RiderAdaptor.java
@@ -1,8 +1,8 @@
 package kitchenpos.order.infrastructure.adaptor;
 
 import java.math.BigDecimal;
-import kitchenpos.order.domain.model.Order;
-import kitchenpos.order.domain.port.RiderPort;
+import kitchenpos.order.domain.order.model.Order;
+import kitchenpos.order.domain.delivery.port.RiderPort;
 import kitchenpos.order.infrastructure.external.KitchenridersClient;
 
 public class RiderAdaptor implements RiderPort {
