+++ src/main/java/kitchenpos/order/infrastructure/persistence/JpaOrderRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.order.infrastructure.persistence;
 
-import kitchenpos.order.domain.model.Order;
-import kitchenpos.order.domain.repository.OrderRepository;
+import kitchenpos.order.domain.order.model.Order;
+import kitchenpos.order.domain.order.repository.OrderRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
