+++ src/main/java/kitchenpos/order/infrastructure/persistence/JpaOrderTableRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.order.infrastructure.persistence;
 
-import kitchenpos.order.domain.model.OrderTable;
-import kitchenpos.order.domain.repository.OrderTableRepository;
+import kitchenpos.order.domain.eatin.model.OrderTable;
+import kitchenpos.order.domain.order.repository.OrderTableRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
