+++ src/main/java/kitchenpos/order/ui/OrderTableRestController.java
@@ -1,7 +1,7 @@
 package kitchenpos.order.ui;
 
-import kitchenpos.order.application.OrderTableService;
-import kitchenpos.order.domain.model.OrderTable;
+import kitchenpos.order.domain.eatin.application.OrderTableService;
+import kitchenpos.order.domain.eatin.model.OrderTable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
