--- src/test/java/kitchenpos/order/infrastructure/persistence/InMemoryOrderTableRepository.java
@@ -1,31 +1,0 @@
-package kitchenpos.order.infrastructure.persistence;
-
-import kitchenpos.order.domain.model.OrderTable;
-import kitchenpos.order.domain.repository.OrderTableRepository;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.UUID;
-
-public class InMemoryOrderTableRepository implements OrderTableRepository {
-    private final Map<UUID, OrderTable> orderTables = new HashMap<>();
-
-    @Override
-    public OrderTable save(final OrderTable orderTable) {
-        orderTables.put(orderTable.getId(), orderTable);
-        return orderTable;
-    }
-
-    @Override
-    public Optional<OrderTable> findById(final UUID id) {
-        return Optional.ofNullable(orderTables.get(id));
-    }
-
-    @Override
-    public List<OrderTable> findAll() {
-        return new ArrayList<>(orderTables.values());
-    }
-}
