+++ src/test/java/kitchenpos/order/order/infrastructure/external/FakeKitchenridersClient.java
@@ -1,0 +1,30 @@
+package kitchenpos.order.order.infrastructure.external;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+import kitchenpos.order.infrastructure.external.KitchenridersClient;
+
+public class FakeKitchenridersClient implements KitchenridersClient {
+    private UUID orderId;
+    private BigDecimal amount;
+    private String deliveryAddress;
+
+    @Override
+    public void requestDelivery(final UUID orderId, final BigDecimal amount, final String deliveryAddress) {
+        this.orderId = orderId;
+        this.amount = amount;
+        this.deliveryAddress = deliveryAddress;
+    }
+
+    public UUID getOrderId() {
+        return orderId;
+    }
+
+    public BigDecimal getAmount() {
+        return amount;
+    }
+
+    public String getDeliveryAddress() {
+        return deliveryAddress;
+    }
+}
