+++ src/test/java/kitchenpos/order/order/infrastructure/persistence/InMemoryOrderRepository.java
@@ -1,0 +1,40 @@
+package kitchenpos.order.order.infrastructure.persistence;
+
+import kitchenpos.order.domain.order.model.Order;
+import kitchenpos.order.domain.order.repository.OrderRepository;
+import kitchenpos.order.domain.order.model.OrderStatus;
+import kitchenpos.order.domain.eatin.model.OrderTable;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
+
+public class InMemoryOrderRepository implements OrderRepository {
+    private final Map<UUID, Order> orders = new HashMap<>();
+
+    @Override
+    public Order save(final Order order) {
+        orders.put(order.getId(), order);
+        return order;
+    }
+
+    @Override
+    public Optional<Order> findById(final UUID id) {
+        return Optional.ofNullable(orders.get(id));
+    }
+
+    @Override
+    public List<Order> findAll() {
+        return new ArrayList<>(orders.values());
+    }
+
+    @Override
+    public boolean existsByOrderTableAndStatusNot(final OrderTable orderTable, final OrderStatus status) {
+        return orders.values()
+            .stream()
+            .anyMatch(order -> order.getOrderTable().equals(orderTable) && order.getStatus() != status);
+    }
+}
