+++ src/test/java/kitchenpos/order/order/infrastructure/persistence/InMemoryOrderTableRepository.java
@@ -1,0 +1,31 @@
+package kitchenpos.order.order.infrastructure.persistence;
+
+import kitchenpos.order.domain.eatin.model.OrderTable;
+import kitchenpos.order.domain.order.repository.OrderTableRepository;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
+
+public class InMemoryOrderTableRepository implements OrderTableRepository {
+    private final Map<UUID, OrderTable> orderTables = new HashMap<>();
+
+    @Override
+    public OrderTable save(final OrderTable orderTable) {
+        orderTables.put(orderTable.getId(), orderTable);
+        return orderTable;
+    }
+
+    @Override
+    public Optional<OrderTable> findById(final UUID id) {
+        return Optional.ofNullable(orderTables.get(id));
+    }
+
+    @Override
+    public List<OrderTable> findAll() {
+        return new ArrayList<>(orderTables.values());
+    }
+}
